LPS-95016 Move RoleFinderTest of the com.liferay.portal.service.persistence package to roles-test module
